Added changes clinton suggested
