+++ card.py
@@ -1,9 +1,11 @@
 
+card_suits = ["Clubs", "Hearts", "Spades", "Diamonds"]
+card_numbers = ["2","3","4","5","6","7","8","9","10", "Ace",
+                "King","Queen","Jack"]
 class Card:
 
-    def __init__(self, number, suit, value):
+    def __init__(self, number, suit):
 
-        self.value = value
         self.suit = suit
         self.number = number
