+++ tests/test_card.py
@@ -1,8 +1,7 @@
 from card import Card
 from deck import Deck
 def test_Card():
-    new_card = Card("King","Spade", 10)
-    assert new_card.value == 10
+    new_card = Card("King","Spade")
     assert new_card.number == "King"
     assert new_card.suit == "Spade"
